Refactoring of ray tracing tools and script from french name to english names.
